integrate / consolidate tests
